+++ changelog.md
@@ -5,6 +5,7 @@
 ### What's Changed
 
 - Added priority of task in automatic plan list.
+- Added check that a task cannot be planned on a day, when the time it would be planned is less than the min_slot_length setting, but the task would normally have more remaining than that. In that case, technically, a task will be planned on the other days instead.
 - Min slot length won't be checked for running tasks now.